requirements.txt
appdirs==1.4.4
APScheduler==3.9.1
beautifulsoup4==4.11.1
black==22.10.0
bs4==0.0.1
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
cssselect==1.1.0
dateparser==1.1.1
demjson3==3.0.5
editdistance==0.6.0
facebook-scraper==0.2.59
fake-useragent==0.1.11
greenlet==1.1.3.post0
idna==3.4
importlib-metadata==5.0.0
lxml==4.9.1
mypy-extensions==0.4.3
parse==1.19.0
pathspec==0.10.1
pip==22.2.2
platformdirs==2.5.2
psycopg2-binary==2.9.4
pyee==8.2.2
pyppeteer==1.0.2
pyquery==1.4.3
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.4
pytz-deprecation-shim==0.1.0.post0
regex==2022.3.2
requests==2.28.1
requests-html==0.10.0
setuptools==63.4.1
six==1.16.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.42
tomli==2.0.1
tqdm==4.64.1
tzdata==2022.5
tzlocal==4.2
urllib3==1.26.12
w3lib==2.0.1
websockets==10.3
wheel==0.37.1
zipp==3.9.0
